[dashboard] refresh ide options on image install
